# openapi.yml
openapi: 3.1.0
info:
  title: Apple Reminders Web API
  description: Apple Reminders 웹 클론 REST API
  version: 0.1.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/lists:
    get:
      tags: [ReminderList]
      summary: 모든 리스트 조회
      operationId: findAllLists
      responses:
        "200":
          description: 리스트 목록 (displayOrder 순)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderListResponse"

    post:
      tags: [ReminderList]
      summary: 리스트 생성
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderListRequest"
      responses:
        "201":
          description: 생성된 리스트
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"
        "400":
          description: 유효성 검증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/lists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [ReminderList]
      summary: 리스트 상세 조회
      operationId: findListById
      responses:
        "200":
          description: 리스트 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [ReminderList]
      summary: 리스트 수정
      operationId: updateList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderListRequest"
      responses:
        "200":
          description: 수정된 리스트
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderListResponse"
        "400":
          description: 유효성 검증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [ReminderList]
      summary: 리스트 삭제
      operationId: deleteList
      responses:
        "204":
          description: 삭제 완료
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/lists/{listId}/reminders:
    parameters:
      - name: listId
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [Reminder]
      summary: 리스트 내 리마인더 조회
      operationId: findRemindersByListId
      responses:
        "200":
          description: 미완료 리마인더 목록 (displayOrder 순)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderResponse"
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Reminder]
      summary: 리마인더 생성
      operationId: createReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderRequest"
      responses:
        "201":
          description: 생성된 리마인더
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"
        "400":
          description: 유효성 검증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 리스트를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/reminders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [Reminder]
      summary: 리마인더 상세 조회
      operationId: findReminderById
      responses:
        "200":
          description: 리마인더 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"
        "404":
          description: 리마인더를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Reminder]
      summary: 리마인더 수정
      operationId: updateReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderRequest"
      responses:
        "200":
          description: 수정된 리마인더
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"
        "400":
          description: 유효성 검증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 리마인더를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Reminder]
      summary: 리마인더 삭제
      operationId: deleteReminder
      responses:
        "204":
          description: 삭제 완료
        "404":
          description: 리마인더를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/reminders/{id}/complete:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    patch:
      tags: [Reminder]
      summary: 완료 토글
      operationId: toggleComplete
      responses:
        "200":
          description: 토글된 리마인더
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"
        "404":
          description: 리마인더를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    ReminderListRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: 리스트 이름
          example: 개인
        color:
          type: string
          description: HEX 색상 코드
          default: "#007AFF"
          example: "#FF3B30"
        icon:
          type: string
          description: 아이콘 식별자
          default: list.bullet
          example: star

    ReminderListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: 개인
        color:
          type: string
          example: "#007AFF"
        icon:
          type: string
          example: list.bullet
        displayOrder:
          type: integer
          example: 0
        reminderCount:
          type: integer
          format: int64
          example: 5
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ReminderRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: 리마인더 제목
          example: 장보기
        notes:
          type: string
          description: 메모
          example: 우유, 빵, 계란

    ReminderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        listId:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: 장보기
        notes:
          type: string
          example: 우유, 빵, 계란
        completed:
          type: boolean
          example: false
        completedAt:
          type: string
          format: date-time
          nullable: true
        displayOrder:
          type: integer
          example: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: 리스트를 찾을 수 없습니다. id=999