# src/main/java/toby/ai/tobyreminder/dto/ReminderRequest.java
package toby.ai.tobyreminder.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Builder;

@Builder
public record ReminderRequest(
        @NotBlank String title,
        String notes
) {
}